fixed type of color_char_to_attr() from byte to int

+++ src/client/externs.h
@@ -987,7 +987,7 @@ extern cptr shortVersion;
 extern void version_build(void);
 extern int find_realm(int book);
 extern char color_attr_to_char(byte a);
-extern byte color_char_to_attr(char c);
+extern int color_char_to_attr(char c);
 extern byte mh_attr(int max);
 extern char *my_strcasestr(const char *big, const char *little);
 extern char *my_strcasestr_skipcol(const char *big, const char *little, byte strict);
